Onyx Data January 2025 Challenge Day 1
This day focused on setting the initial idea for the dashboard, looking at the requirements and seeing what we could do with the data. I started by dividing the problem after looking at the brief given by the challenge, and ended on the idea of having 2 sides to the dashboard, "Sales" and "Customer".

Started work on the "Sales" portion of the dashboard.